Simplified DAOs and Models

+++ src/main/resources/db/migration/V1__tableCreation.sql
@@ -1,13 +1,13 @@
 create table user
 (
-    id       int primary key auto_increment,
+    id       bigint primary key auto_increment,
     username varchar not null unique,
     password varchar not null
 );
 
 create table message
 (
-    id        int primary key auto_increment,
+    id        bigint primary key auto_increment,
     text      varchar  not null,
     send_to   int      not null,
     sent_from int      not null,
